Relocate pairs and policy under iam dir
