+++ components/automate-gateway/gateway/middleware/authv2/authv2.go
@@ -10,8 +10,8 @@ import (
 	"google.golang.org/grpc/status"
 
 	authz "github.com/chef/automate/api/interservice/authz/v2"
-	"github.com/chef/automate/components/automate-gateway/api/authz/pairs"
-	"github.com/chef/automate/components/automate-gateway/api/authz/policy"
+	"github.com/chef/automate/components/automate-gateway/api/iam/v2/pairs"
+	"github.com/chef/automate/components/automate-gateway/api/iam/v2/policy"
 	"github.com/chef/automate/components/automate-gateway/gateway/middleware"
 	"github.com/chef/automate/lib/grpc/auth_context"
 )
