+++ components/automate-gateway/gateway/services.go
@@ -38,12 +38,12 @@ import (
 	"github.com/chef/automate/api/interservice/compliance/reporting"
 	deploy_api "github.com/chef/automate/api/interservice/deployment"
 	swagger "github.com/chef/automate/components/automate-gateway/api"
-	policy "github.com/chef/automate/components/automate-gateway/api/authz/policy"
 	pb_deployment "github.com/chef/automate/components/automate-gateway/api/deployment"
 	pb_eventfeed "github.com/chef/automate/components/automate-gateway/api/event_feed"
 	pb_gateway "github.com/chef/automate/components/automate-gateway/api/gateway"
 	pb_authz "github.com/chef/automate/components/automate-gateway/api/iam/v2"
 	pb_iam "github.com/chef/automate/components/automate-gateway/api/iam/v2"
+	policy "github.com/chef/automate/components/automate-gateway/api/iam/v2/policy"
 	pb_legacy "github.com/chef/automate/components/automate-gateway/api/legacy"
 	pb_license "github.com/chef/automate/components/automate-gateway/api/license"
 	pb_notifications "github.com/chef/automate/components/automate-gateway/api/notifications"
