+++ components/automate-gateway/pkg/authorizer/authorizer.go
@@ -4,7 +4,7 @@ import (
 	"context"
 	"fmt"
 
-	"github.com/chef/automate/components/automate-gateway/api/authz/pairs"
+	"github.com/chef/automate/components/automate-gateway/api/iam/v2/pairs"
 	"github.com/chef/automate/components/automate-gateway/gateway/middleware"
 )
